Canonical code form
