+++ middleware/header.go
@@ -17,7 +17,7 @@ func setRespHeaders() macaron.Handler {
 			ctx.Resp.Header().Set("X-Docker-Registry-Config", setting.RunMode)
 			ctx.Resp.Header().Set("X-Docker-Endpoints", setting.Domains)
 		} else if flag := strings.Contains(ctx.Req.RequestURI, "/v2/"); flag == true {
-			ctx.Resp.Header().Set("Content-Type", "text/plain; charset=utf-8")
+			//ctx.Resp.Header().Set("Content-Type", "text/plain; charset=utf-8")
 			//ctx.Resp.Header().Set("WWW-Authenticate", fmt.Sprintf("Basic realm=\"%v\"", setting.Domains))
 			ctx.Resp.Header().Set("Docker-Distribution-Api-Version", setting.DistributionVersion)
 		} else {
